+++ councilmatic/settings_deployment.py.example
@@ -92,3 +92,8 @@ LOGGING = {
         },
     },
 }
+
+# Google calendar API
+SCOPES = 'https://www.googleapis.com/auth/calendar' # for read/write access
+CLIENT_SECRET_FILE = '' # the name of your Google-generated json file
+APPLICATION_NAME = 'NYC Councilmatic'
